+++ blink/fds.h
@@ -1,8 +1,10 @@
 #ifndef BLINK_FDS_H_
 #define BLINK_FDS_H_
+#include <limits.h>
 #include <poll.h>
-#include <stdint.h>
 #include <sys/uio.h>
+
+#include "blink/types.h"
 
 struct MachineFdClosed {
   int fd;
@@ -13,7 +15,7 @@ struct MachineFdCb {
   int (*close)(int);
   ssize_t (*readv)(int, const struct iovec *, int);
   ssize_t (*writev)(int, const struct iovec *, int);
-  int (*ioctl)(int, int, ...);
+  int (*ioctl)(int, unsigned long, ...);
   int (*poll)(struct pollfd *, nfds_t, int);
 };
